Initial commit - simple menu, project layout, basic networking
Created the basic layout for the project, and created a simple title
menu, along with a button class. Also added the precursor code for
re-scaling the window. Added the 'Galdeano' font. Added basic networking
functionality, including the ability to join the multiplayer queue, and
check when a game has been found.