Added game joining system
Added the basic handshake system to join the game, including getting
which player is going first. Added a simple screen to tell the player if
they are going first or second. Created a new 'guiItems' file, and moved
the blackFade class into it. Added the ability to use a radial gradient
in the blackFade class. Added the ability to pause, and moved drawing to
a new function in all game items. Also added the mana cost of a card.