Massively improved performance.
Only portions of the display that were edited are updated. All images
are converted. The game is now profiled, and the profile output is saved
in 'logs/Calltime Dump'. A new file named 'renderer' contains the
display refresh function and the display changelist function. Average
FPS in fullscreen mode is now 60.