Added network event infrastructure and a debug monitor
Can now only drag cards if it is the player's turn. Card placement is
now synced with the server. Also, all cards now have a unique identifier
independent of the card's name. Added a debug window in the top left
corner that shows the mouse position and fps. Also fixed button
centring, and added a separate draw button to the button class. Added an
end turn button.