Card movements now synchronise
Cards are now assigned a unid, and a player. Only friendly cards can now
be used. Moved incoming network event handing from the network thread to
the game loop. The opponent's unid is now fetched at the start. Moving a
card will now send a network event that can be read and executed from
the client.