Added a turn timer
Added a turn timer that slowly fades in and shows how much time is left
in the turn. When the time is up, the turn switches to the opponent.
Also, tiles are now coloured based on who owns the card on them - red
for opponent, green for player. Moved the block of code that implements
the events on the network changes list to its own function in the
network module.