Fixed exceptions not being printed, and added a crash report screen
Added a new crash handling system, which includes saving a crash file,
and displaying the traceback in a new 'crash' state. Can now specify a
reason for why the client is disconnecting from the game (currently not
used). Fixed casting spells not re-ordering the hand. Added a default
font (for when a font = None).