Added login screen
Added a login screen with a fire effect in the background. Added two new
fonts for the title. Made lots of tweaks to the text input box to make
it work properly, added some new options to it, added some
documentation, added the ability to hide the contents for passwords.
Also added an animation class, to use to animate sprite sheets.