Added account stat system, updating of stats on game finish
Added functions for updating and getting account stats. On game end,
wins and losses are updated, and a change in competitive points is
calculated. Also added username and password saving. Added rect
calculation to text. Made in-game player name the username. Added the
loading and saving of a pickled data file (that at the moment contains
the username, password and game version).