Added options menu and infrastructure
Added the options state. Added the Option class, for adding option menu
objects. Added the InfoBubble class, for explaining what each option
does. Added the info icon. Can now change the frequency at which the
server is pinged in the options. More options are now saved and loaded
in the data file.