Added campaign map
Added the campaign state, which contains a map. Added a flag class and
flag assets. Flags represent levels, and can be hovered and clicked.
Added the level overview class, which contains the level's name,
description and play button - new fonts and background. Added groundwork
for playing offline with an ai opponent.