Added the AI for the single player campaign
The AI works by assigning each enemy card, and the enemy castle, a rank.
The rank is based on the card's cost, attack, health, % health, and
distance from the AI castle. Each AI card then makes a decision of what
to do based on those rankings, and where that card is. The AIs decision
making can be altered by adjusting various 'weight' variables.